# angular/src/app/services/models/appointment-response.ts
/* tslint:disable */
/* eslint-disable */
export interface AppointmentResponse {
  available?: boolean;
  bookedStudent?: string;
  date?: string;
  eligibleStudents?: Array<string>;
  endTime?: string;
  id?: number;
  instructor?: string;
  startTime?: string;
}